NetworkController angular methods
Extrack common behaviour of cloud controller and make it
available for NetworkController
