+++ app/controllers/ems_network_controller.rb
@@ -1,5 +1,7 @@
 class EmsNetworkController < ApplicationController
   include EmsCommon
+  include Mixins::EmsCommonAngular
+  include Mixins::GenericSessionMixin
 
   before_action :check_privileges
   before_action :get_session_data
@@ -14,7 +16,11 @@ class EmsNetworkController < ApplicationController
     @table_name ||= "ems_network"
   end
 
-  def index
-    redirect_to :action => 'show_list'
+  def ems_path(*args)
+    ems_network_path(*args)
+  end
+
+  def ems_network_form_fields
+    ems_form_fields
   end
 end
