+++ app/controllers/mixins/generic_session_mixin.rb
@@ -2,8 +2,9 @@ module Mixins
   module GenericSessionMixin
     private
     def get_session_data
-      @title      = self.class.title
+      @title      = ui_lookup(:tables => self.class.table_name)
       @layout     = self.class.table_name
+      @table_name = request.parameters[:controller]
       @lastaction = session["#{self.class.table_name}_lastaction".to_sym]
       @display    = session["#{self.class.table_name}_display".to_sym]
       @filters    = session["#{self.class.table_name}_filters".to_sym]
