+++ app/views/ems_network/edit.html.haml
@@ -1,11 +1,11 @@
 = form_for(@ems,
-           :url    => ems_cloud_path(@ems),
+           :url    => ems_network_path(@ems),
            :method => :patch,
            :html   => {"ng-controller"   => "emsCommonFormController",
                        "name"            => "angularForm",
                        "ng-show"         => "afterGet",
-                       "update-url"      => "#{ems_cloud_path(@ems)}",
-                       "form-fields-url" => "/#{controller_name}/ems_cloud_form_fields/",
+                       "update-url"      => "#{ems_network_path(@ems)}",
+                       "form-fields-url" => "/#{controller_name}/ems_network_form_fields/",
                        "novalidate"      => true}) do |f|
   %input{:type => 'hidden', :id => "form_id", :value => "##{f.options[:html][:id]}"}
   %input{:type => 'hidden', :id => "button_name", :name => "button", :value => "save"}
