+++ app/views/ems_network/new.html.haml
@@ -1,4 +1,4 @@
-- url = @ems.persisted? ? ems_clouds_path(@ems) : ems_clouds_path
+- url = @ems.persisted? ? ems_network_path(@ems) : ems_network_path
 = form_for(@ems,
            :url    => url,
            :method => :post,
@@ -6,7 +6,7 @@
                        "name"            => "angularForm",
                        "ng-show"         => "afterGet",
                        "create-url"      => "#{url}",
-                       "form-fields-url" => "/#{controller_name}/ems_cloud_form_fields/",
+                       "form-fields-url" => "/#{controller_name}/ems_network_form_fields/",
                        "novalidate"      => true}) do |f|
   %input{:type => 'hidden', :id => "form_id", :value => "##{f.options[:html][:id]}"}
   %input{:type => 'hidden', :id => "button_name", :name => "button", :value => "add"}
